test: fold workflow file action host suite
