+++ src/test-kernel/controllers/ProgressWorkflowFileActionsController.vitest.ts
@@ -7,9 +7,47 @@ import {
   type ProgressWorkflowFileActionsControllerDeps,
 } from '@controllers/progressView/ProgressWorkflowFileActionsController';
 
+type LogEntry = { message: string; error: unknown };
+
+type RecordingHost = ProgressWorkflowFileActionsControllerDeps['host'] & {
+  infos: string[];
+  errors: string[];
+  logs: LogEntry[];
+};
+
+type RecordingDeps = ProgressWorkflowFileActionsControllerDeps & {
+  host: RecordingHost;
+};
+
 function createDeps(
   overrides: Partial<ProgressWorkflowFileActionsControllerDeps['host']>,
-): ProgressWorkflowFileActionsControllerDeps {
+): RecordingDeps {
+  const infos: string[] = [];
+  const errors: string[] = [];
+  const logs: LogEntry[] = [];
+  const host: RecordingHost = {
+    infos,
+    errors,
+    logs,
+    compareFiles: async () => {},
+    acceptEditedFile: async () => {},
+    mergeFile: async () => {},
+    latexdiffFile: async () => {},
+    openDirectory: async () => {},
+    openLabel: async () => true,
+    readFile: async () => '',
+    showInfo: async (message) => {
+      infos.push(message);
+    },
+    showError: async (message) => {
+      errors.push(message);
+    },
+    logError: (message, error) => {
+      logs.push({ message, error });
+    },
+    ...overrides,
+  };
+
   return {
     state: {
       getActiveStream: () => '',
@@ -17,18 +55,7 @@ function createDeps(
       getOutputFiles: () => ({}),
       getAgentModel: () => undefined,
     },
-    host: {
-      compareFiles: async () => {},
-      acceptEditedFile: async () => {},
-      mergeFile: async () => {},
-      latexdiffFile: async () => {},
-      openDirectory: async () => {},
-      openLabel: async () => true,
-      readFile: async () => '',
-      showInfo: async () => {},
-      showError: async () => {},
-      ...overrides,
-    },
+    host,
     sendFollowUp: async () => {},
   };
 }
@@ -57,5 +84,68 @@ describe('ProgressWorkflowFileActionsController', () => {
       ['/workspace/output-r0.tex', '/workspace/output-r1.tex'],
     ]);
     expect(latexdiffs).toEqual([]);
+  });
+
+  it('shows one fallback message when a host cannot open a label', async () => {
+    const deps = createDeps({
+      openLabel: async () => false,
+    });
+    const controller = new ProgressWorkflowFileActionsController(deps);
+
+    await controller.openLabel('missing-label');
+
+    expect(deps.host.infos).toEqual(['Label "missing-label" not found.']);
+  });
+
+  it('reports task storage open failures', async () => {
+    const failure = new Error('cannot reveal folder');
+    const deps = createDeps({
+      openDirectory: async () => {
+        throw failure;
+      },
+    });
+    deps.state.getExecutionId = () => 'run-1';
+    const controller = new ProgressWorkflowFileActionsController(deps);
+
+    await controller.openTaskStorage('toolUse');
+
+    expect(deps.host.errors).toEqual([
+      'Failed to open task storage folder: cannot reveal folder',
+    ]);
+    expect(deps.host.logs).toEqual([
+      { message: 'Failed to open task storage folder', error: failure },
+    ]);
+  });
+
+  it('keeps the accept follow-up backup when the host cancels acceptance', async () => {
+    const followUps: string[] = [];
+    const acceptResults = [false, true];
+    const readResults = [
+      'original model output',
+      'user edited output',
+      'user edited output',
+    ];
+    const deps = createDeps({
+      acceptEditedFile: async () => acceptResults.shift(),
+      readFile: async () => readResults.shift() ?? '',
+    });
+    deps.state.getActiveStream = () => 'workflow';
+    deps.sendFollowUp = async (_stream, text) => {
+      followUps.push(text);
+    };
+    const controller = new ProgressWorkflowFileActionsController(deps);
+
+    await controller.compareOriginal(
+      '/workspace/edited.tex',
+      '/workspace/base.tex',
+    );
+    await controller.acceptFile('/workspace/edited.tex', '/workspace/base.tex');
+
+    expect(followUps).toEqual([]);
+
+    await controller.acceptFile('/workspace/edited.tex', '/workspace/base.tex');
+
+    expect(followUps).toHaveLength(1);
+    expect(followUps[0]).toMatch(/edited\.tex/);
   });
 });
